\name{makePlot.gene.assignment.rates}
\docType{methods}
\alias{makePlot.gene.assignment.rates}
\title{
   Gene assignment rate plot
}
\description{
   Plots the rate at which reads are assigned into various 
   categories.
}
\usage{
makePlot.gene.assignment.rates(plotter,  debugMode, \dots)
                         
}
\arguments{
  \item{plotter}{
    A QoRT_Plotter object.
  }
  \item{debugMode}{
    Logical. If TRUE, debugging data will be printed to the console.
  }
  \item{\dots}{ 
    Arguments to be passed to methods, such as
    \link{graphical parameters} (see \code{\link{par}}).
  }
}
\details{
   TODO!
}

\seealso{

}