\name{makePlot.splice.junction.event.rates}
\docType{methods}
\alias{makePlot.splice.junction.event.rates}
\alias{makePlot.splice.junction.event.rates.split.by.type}
\title{
   Plot Splice Junction Event Rates
}
\description{
   Plots the rates at which splice junctions occur.
}
\usage{
   makePlot.splice.junction.event.rates(plotter, 
                                    calc.ratePerRead = TRUE, 
                                    calc.proportion = FALSE, 
                                    high.low.cutoff = 4, 
                                    debugMode = DEFAULTDEBUGMODE, 
                                    ...)
   
   makePlot.splice.junction.event.rates.split.by.type(plotter, 
                                    calc.rate = TRUE, 
                                    high.low.cutoff = 4, 
                                    debugMode = DEFAULTDEBUGMODE, 
                                    ...)
}
\arguments{
  \item{plotter}{
    A QoRT_Plotter object.
  }
  \item{calc.ratePerRead}{
    TODO!!!!!
  }
  \item{calc.proportion}{
    TODO!!!!!
  }
  \item{high.low.cutoff}{
    TODO!!!!!
  }
  \item{debugMode}{
    Activates debug mode, which causes more verbose reporting.
  }
  
  \item{calc.rate}{
    TODO!!!!!
  }
  \item{\dots}{ 
    Arguments to be passed to methods, such as
    \link{graphical parameters} (see \code{\link{par}}).
  }
}
\details{
   TODO!
}

\seealso{

}