\name{get.summary.table}
\docType{methods}
\alias{get.summary.table}
\title{
   Get summary data table
}
\description{
   Retrieves and compiles a summary data table.
}
\usage{
  get.summary.table(res, debugMode)
}
\arguments{
  \item{res}{
    A QoRT_QC_Results object.
  }
  \item{debugMode}{
    Logical. If TRUE, debugging data will be printed to the console.
  }
}
\details{
  Returns summary data in table form.
}

\seealso{

}