\name{makePlot.dropped.rates}
\docType{methods}
\alias{makePlot.dropped.rates}
\title{
   Read Drop Plot
}
\description{
   Plots the rates at which reads are dropped from
   analysis for various causes.
}
\usage{
makePlot.dropped.rates(plotter, debugMode, \dots)
}
\arguments{
  \item{plotter}{
    A QoRT_Plotter object.
  }
  \item{debugMode}{
    Logical. If TRUE, debugging data will be printed to the console.
  }
  \item{\dots}{ 
    Arguments to be passed to methods, such as
    \link{graphical parameters} (see \code{\link{par}}).
  }
}
\details{
   TODO!
}

\seealso{

}